Additional Changes
So I added comments to my files according to what I have researched on
the internet. Please let me know if there are any concerns or glaring
errors. I might or might not make anymore additional changes to these,
but if I do, it will be before the Thursday deadline. Otherwise, I guess
this will be my final final.